refactored coin line graph into reusable CoinLineGraph component
